Refactor anchorEl using useStandardMenuList hook
